First half of Event refactor
Move helper functions, types, and constants out of CreateModifyEvent
